06042023-1728

+++ Business/Abstract/ICartService.cs
@@ -12,9 +12,9 @@ namespace Business.Abstract
 {
     public interface ICartService
     {
-        IResult Add(int productId,int quantity,string token);
+        IResult Add(int productId,short quantity,string token);
         IResult Update(Cart cart);
-        IResult Delete(int productId, int quantity, string token);
+        IResult Delete(int productId, short quantity, string token);
         IDataResult<List<ProductPreviewDto>> GetList(string token);
         IDataResult<Cart> GetById(int cartId);
         IDataResult<Cart> GetByUserId(int userId);
